KPMP-148(rlreamy/cgates/zwright): Moved the mysql classes into a deprecated folder. Added new mappings for mongo
