+++ src/main/java/org/kpmp/GenerateUniversalIds.java
@@ -2,9 +2,9 @@ package org.kpmp;
 
 import java.util.List;
 
-import org.kpmp.dao.FileSubmission;
-import org.kpmp.dao.UploadPackage;
-import org.kpmp.upload.UploadPackageRepository;
+import org.kpmp.dao.deprecated.FileSubmission;
+import org.kpmp.dao.deprecated.UploadPackage;
+import org.kpmp.upload.UploadPackageMySQLRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
@@ -16,11 +16,11 @@ import org.springframework.context.annotation.ComponentScan;
 @ComponentScan(basePackages = { "org.kpmp" })
 public class GenerateUniversalIds implements CommandLineRunner {
 
-	private UploadPackageRepository packageRepository;
+	private UploadPackageMySQLRepository packageRepository;
 	private UniversalIdGenerator uuidGenerator;
 
 	@Autowired
-	public GenerateUniversalIds(UploadPackageRepository packageRepository, UniversalIdGenerator uuidGenerator) {
+	public GenerateUniversalIds(UploadPackageMySQLRepository packageRepository, UniversalIdGenerator uuidGenerator) {
 		this.packageRepository = packageRepository;
 		this.uuidGenerator = uuidGenerator;
 	}
