+++ src/main/java/org/kpmp/dao/deprecated/Protocol.java
@@ -1,4 +1,4 @@
-package org.kpmp.dao;
+package org.kpmp.dao.deprecated;
 
 import java.util.List;
 
@@ -10,16 +10,16 @@ import javax.persistence.OneToMany;
 import javax.persistence.Table;
 
 @Entity
-@Table(name = "package_type")
-public class PackageType {
+@Table(name = "protocol")
+public class Protocol {
 
 	@Id
 	@Column(name = "id")
 	private int id;
-	@Column(name = "package_type")
-	private String packageType;
+	@Column(name = "protocol")
+	private String protocol;
 
-	@OneToMany(cascade = CascadeType.ALL, mappedBy = "packageType")
+	@OneToMany(cascade = CascadeType.ALL, mappedBy = "protocol")
 	private List<UploadPackage> uploadPackages;
 
 	public int getId() {
@@ -30,12 +30,12 @@ public class PackageType {
 		this.id = id;
 	}
 
-	public String getPackageType() {
-		return packageType;
+	public String getProtocol() {
+		return protocol;
 	}
 
-	public void setPackageType(String packageType) {
-		this.packageType = packageType;
+	public void setProtocol(String protocol) {
+		this.protocol = protocol;
 	}
 
 	public List<UploadPackage> getUploadPackages() {
