+++ src/main/java/org/kpmp/upload/InstitutionRepository.java
@@ -1,6 +1,6 @@
 package org.kpmp.upload;
 
-import org.kpmp.dao.InstitutionDemographics;
+import org.kpmp.dao.deprecated.InstitutionDemographics;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
