+++ src/main/java/org/kpmp/upload/PackageTypeOtherRepository.java
@@ -2,7 +2,7 @@ package org.kpmp.upload;
 
 import javax.transaction.Transactional;
 
-import org.kpmp.dao.PackageTypeOther;
+import org.kpmp.dao.deprecated.PackageTypeOther;
 import org.springframework.data.repository.CrudRepository;
 
 public interface PackageTypeOtherRepository extends CrudRepository<PackageTypeOther, Integer> {
