+++ src/main/java/org/kpmp/upload/PackageTypeRepository.java
@@ -1,6 +1,6 @@
 package org.kpmp.upload;
 
-import org.kpmp.dao.PackageType;
+import org.kpmp.dao.deprecated.PackageType;
 import org.springframework.data.repository.CrudRepository;
 
 public interface PackageTypeRepository extends CrudRepository<PackageType, Integer> {
