+++ src/main/java/org/kpmp/upload/ProtocolRepository.java
@@ -1,6 +1,6 @@
 package org.kpmp.upload;
 
-import org.kpmp.dao.Protocol;
+import org.kpmp.dao.deprecated.Protocol;
 import org.springframework.data.repository.CrudRepository;
 
 public interface ProtocolRepository extends CrudRepository<Protocol, Integer> {
