+++ src/main/java/org/kpmp/upload/SubmitterRepository.java
@@ -2,7 +2,7 @@ package org.kpmp.upload;
 
 import javax.transaction.Transactional;
 
-import org.kpmp.dao.SubmitterDemographics;
+import org.kpmp.dao.deprecated.SubmitterDemographics;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
