+++ src/main/java/org/kpmp/upload/UploadService.java
@@ -5,22 +5,22 @@ import java.io.IOException;
 import java.util.Date;
 
 import org.kpmp.UniversalIdGenerator;
-import org.kpmp.dao.FileMetadataEntries;
-import org.kpmp.dao.FileSubmission;
-import org.kpmp.dao.FileSubmissionsRepository;
-import org.kpmp.dao.InstitutionDemographics;
-import org.kpmp.dao.PackageType;
-import org.kpmp.dao.PackageTypeOther;
-import org.kpmp.dao.Protocol;
-import org.kpmp.dao.SubmitterDemographics;
-import org.kpmp.dao.UploadPackage;
+import org.kpmp.dao.deprecated.FileMetadataEntries;
+import org.kpmp.dao.deprecated.FileSubmission;
+import org.kpmp.dao.deprecated.FileSubmissionsRepository;
+import org.kpmp.dao.deprecated.InstitutionDemographics;
+import org.kpmp.dao.deprecated.PackageType;
+import org.kpmp.dao.deprecated.PackageTypeOther;
+import org.kpmp.dao.deprecated.Protocol;
+import org.kpmp.dao.deprecated.SubmitterDemographics;
+import org.kpmp.dao.deprecated.UploadPackage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
 public class UploadService {
 
-	private UploadPackageRepository uploadPackageRepository;
+	private UploadPackageMySQLRepository uploadPackageRepository;
 	private FileSubmissionsRepository fileSubmissionsRepository;
 	private SubmitterRepository submitterRepository;
 	private InstitutionRepository institutionRepository;
@@ -31,7 +31,7 @@ public class UploadService {
 	private ProtocolRepository protocolRepository;
 
 	@Autowired
-	public UploadService(UploadPackageRepository uploadPackageRepository,
+	public UploadService(UploadPackageMySQLRepository uploadPackageRepository,
 			FileSubmissionsRepository fileSubmissionsRepository, SubmitterRepository submitterRepository,
 			InstitutionRepository institutionRepository, FileMetadataRepository fileMetadataRepository,
 			PackageTypeRepository packageTypeRepository, PackageTypeOtherRepository packageTypeOtherRepository,
