+++ src/main/resources/application.properties
@@ -28,3 +28,5 @@ metadata.file.name=metadata.json
 spring.servlet.multipart.max-file-size=102400MB
 spring.servlet.multipart.max-request-size=102400MB
 spring.http.multipart.enabled=false
+
+spring.data.mongodb.uri=mongodb://mongodb:27017/dataLake
