+++ src/test/java/org/kpmp/dao/FileFormatsTest.java
@@ -8,6 +8,8 @@ import java.util.List;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+import org.kpmp.dao.deprecated.FileFormats;
+import org.kpmp.dao.deprecated.FileSubmission;
 
 public class FileFormatsTest {
