+++ src/test/java/org/kpmp/dao/InstitutionDemographicsTest.java
@@ -8,6 +8,8 @@ import java.util.List;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+import org.kpmp.dao.deprecated.FileSubmission;
+import org.kpmp.dao.deprecated.InstitutionDemographics;
 
 public class InstitutionDemographicsTest {
