+++ src/test/java/org/kpmp/dao/SubmitterDemographicsTest.java
@@ -9,6 +9,8 @@ import java.util.List;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+import org.kpmp.dao.deprecated.FileSubmission;
+import org.kpmp.dao.deprecated.SubmitterDemographics;
 import org.kpmp.upload.PackageInformation;
 
 public class SubmitterDemographicsTest {
