+++ src/test/java/org/kpmp/view/FileUploadTest.java
@@ -9,11 +9,11 @@ import java.util.Date;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import org.kpmp.dao.FileSubmission;
-import org.kpmp.dao.InstitutionDemographics;
-import org.kpmp.dao.PackageType;
-import org.kpmp.dao.SubmitterDemographics;
-import org.kpmp.dao.UploadPackage;
+import org.kpmp.dao.deprecated.FileSubmission;
+import org.kpmp.dao.deprecated.InstitutionDemographics;
+import org.kpmp.dao.deprecated.PackageType;
+import org.kpmp.dao.deprecated.SubmitterDemographics;
+import org.kpmp.dao.deprecated.UploadPackage;
 
 public class FileUploadTest {
